# contracts/doorstep/openapi.yaml
openapi: 3.0.3
info:
  title: Doorstep Home Services API
  version: 1.0.0
  description: >-
    Authoritative REST contract for Doorstep (Urban Company style home
    services): the customer surface (Momentum Android app and postbook-ui
    /doorstep), the professional surface (/pro, the doorstep_pro Android app)
    and the admin-internal family (/internal/admin, admin-service tokens only).


    Served by doorstep-service (port 8122, schema `doorstep`) behind the API
    gateway prefix /v1/doorstep (dormant gate DOORSTEP_PUBLIC_ENABLED, pilot
    allowlist DOORSTEP_PILOT_USER_IDS). Pilot city Hyderabad (`HYD`).


    Conventions (binding for every lane):

    - Every response is the shared envelope: success `{"data": <payload>,
      "meta": {"request_id": "..."}}`, failure `{"error": {"code", "message",
      "details"?}, "meta": {...}}`. Schemas below describe `data` unless they
      are named `*Envelope` or `ErrorEnvelope`.

    - Money is integer paise in fields ending `_paise`. Customer prices are
      GST-inclusive; the split is shown as `taxable_paise` + `tax_paise`.

    - Error codes are stable UPPER_SNAKE strings (see `x-doorstep-error-codes`).
      Clients branch on `error.code`, never on `message`.

    - Timestamps are RFC 3339 UTC. Dates are `YYYY-MM-DD` in the city's time
      zone (Asia/Kolkata). Times of day are `HH:MM`.

    - Optional values are emitted as explicit `null`, never omitted, on every
...

...
      gateway refuses /internal/ from the edge.


...


...
servers:
...
security:
  - gatewayIdentity: []
...
paths:
...
  /v1/doorstep/catalogue:
    get:
      summary: Categories offered in a city
      operationId: getCatalogue
      tags: [Catalogue]
      security: []
      x-lane: A1
      parameters:
        - $ref: '#/components/parameters/CityQuery'
      responses:
        '200':
          description: Active categories of a family the tax computer prices (B1 families wait for the tax lane) with at least one active service; starting_price_paise is the lowest bookable professional price (null while nobody offers one). Golden bytes catalogue_get_200.json.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Catalogue'}
        '404':
          $ref: '#/components/responses/Error'
  /v1/doorstep/categories/{slug}:
    get:
      summary: One category with its services
      operationId: getCategory
      tags: [Catalogue]
      security: []
      x-lane: A1
      parameters:
        - name: slug
          in: path
          required: true
          schema: {type: string, example: salon-women}
        - $ref: '#/components/parameters/CityQuery'
      responses:
        '200':
          description: Golden bytes category_get_200.json.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CategoryPage'}
        '404':
          description: DOORSTEP_CITY_NOT_FOUND or DOORSTEP_CATEGORY_NOT_FOUND
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /v1/doorstep/services/{id}:
    get:
      summary: Service detail (the menu) with options, units, add-on groups, suggested and from prices
      operationId: getService
      tags: [Catalogue]
      security: []
      x-lane: A1
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/CityQuery'
      responses:
        '200':
          description: Active options (with unit, the city's suggested price and the lowest bookable professional price) and active add-ons; prices are the professionals', picked through GET /services/{id}/professionals. Golden bytes service_get_200.json.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ServicePage'}
        '404':
          description: DOORSTEP_CITY_NOT_FOUND, DOORSTEP_SERVICE_NOT_FOUND
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '422':
          description: DOORSTEP_SERVICE_NOT_AVAILABLE (no active option, or details.reason tax_category_pending)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /v1/doorstep/serviceability:
    post:
      summary: Is a point inside an active service zone?
      operationId: checkServiceability
      tags: [Catalogue]
      security: []
      x-lane: A1
      description: >-
        Point-in-zone via PostGIS ST_Covers on the zone boundary (a point on
        the edge is inside). When zones touch, the lowest (sort by name, id)
        zone wins. Always 200; `serviceable=false` carries reason
        OUTSIDE_SERVICE_AREA. Golden bytes serviceability_in_200.json,
        serviceability_out_200.json.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ServiceabilityRequest'}
      responses:
        '200':
          description: Serviceability answer
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Serviceability'}
        '400':
          $ref: '#/components/responses/Error'
  /v1/doorstep/share/{token}:
    get:
      summary: Shared booking status (trusted contact view)
      operationId: getSharedBooking
      tags: [Safety]
      security: []
      x-lane: A5
      parameters:
        - name: token
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SharedBookingView'}
          description: Locality, status, professional first name and photo, ETA. Never the full address.
        '404':
          description: DOORSTEP_SHARE_TOKEN_INVALID
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}

  # ===================================================================
  # Customer (gateway identity)
  # ===================================================================
  /v1/doorstep/quotes:
    post:
      summary: Price a service selection with the picked professional (GST-inclusive)
      operationId: createQuote
      tags: [Customer]
      x-lane: A1
      description: >-
        B1: pro_id (required) is the professional the customer picked; every
        line is priced with that professional's approved, live price (no city
        price ever fills a gap: DOORSTEP_PRICE_UNAVAILABLE, details.item_ids)
        and price_id is that price row. Validates the option (active, of this
        service), the quantity (1..max_quantity) and the add-ons with the food
        add-on group rules (each group: count >= max(min_select, is_required ?
        1 : 0) and count <= max_select; no duplicates; add-on of this service,
        active). Prices are GST-inclusive; tax is split by
        the TaxComputer (provisional until shared/gst lands; the invoice is
        recomputed at completion with the actual professional). Expires after
        15 minutes. Golden bytes quote_post_201.json, quote_post_422_*.json.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/QuoteRequest'}
      responses:
        '201':
          description: Quote created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Quote'}
        '400':
          $ref: '#/components/responses/Error'
        '404':
          description: DOORSTEP_SERVICE_NOT_FOUND
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '422':
          description: DOORSTEP_OUTSIDE_SERVICE_AREA, DOORSTEP_OPTION_INVALID, DOORSTEP_QUANTITY_INVALID, DOORSTEP_ADDON_INVALID, DOORSTEP_SERVICE_NOT_AVAILABLE, DOORSTEP_PRICE_UNAVAILABLE (quote_post_422_price_unavailable.json; quote_post_400_pro_required.json without pro_id)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /v1/doorstep/quotes/{id}:
    get:
      summary: Read one of my quotes
      operationId: getQuote
      tags: [Customer]
      x-lane: A1
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Quote'}
          description: The quote (status open, expired or consumed)
        '404':
          description: DOORSTEP_QUOTE_NOT_FOUND (also for another user's quote)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
  /v1/doorstep/webhooks/background-check/{provider}:
    post:
      summary: Background-check vendor callback (AuthBridge / IDfy / OnGrid, later)
      operationId: backgroundCheckWebhook
      tags: [Webhook]
      security: []
      x-lane: A2
      description: Verified by the provider's HMAC signature header; unknown provider -> 404; bad signature -> 401. No provider is enabled at launch.
      parameters:
        - name: provider
          in: path
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, additionalProperties: true}
      responses:
        '204':
          description: Applied (idempotent on the provider event id)

  # ===================================================================
  # Admin-internal (admin-service tokens only)
  # ===================================================================
...
  /v1/doorstep/internal/admin/professionals/{id}/tax-registration:
    get:
      summary: Read the reviewed GST registration (sensitive identity)
      operationId: adminTaxRegistration
      tags: [AdminProfessional]
      x-lane: A6
      x-permission: doorstep:pros.approve
      x-step-up: true
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Registration; private and no-store
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {$ref: '#/components/schemas/TaxRegistration'}
    post:
      summary: Record a manually verified registration, or remove it; refuses active bookings
      operationId: adminSetTaxRegistration
      tags: [AdminProfessional]
      x-lane: A6
      x-permission: doorstep:pros.approve
      x-step-up: true
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [gstin, reason]
              additionalProperties: false
              properties:
                gstin: {type: string, description: 'Blank removes the registration; otherwise a checksum-valid GSTIN'}
                verified: {type: boolean, description: 'Required true for nonblank GSTIN after checking active registration and ownership'}
                reason: {type: string, minLength: 10, maxLength: 1000}
      responses:
        '200':
          description: Recorded and audited atomically; does not approve the professional
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {$ref: '#/components/schemas/TaxRegistration'}
        '409':
          description: Active bookings prevent tax treatment changes
...
  /v1/doorstep/internal/admin/incidents/{id}/acknowledge:
    post:
      summary: Acknowledge an incident
      operationId: adminAcknowledgeIncident
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:incidents.act
      x-lane: A6
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Incident
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Incident'}
  /v1/doorstep/internal/admin/incidents/{id}/resolve:
    post:
      summary: Resolve (lifts an incident auto-suspension only when lift_suspension=true)
      operationId: adminResolveIncident
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:incidents.act
      x-lane: A6
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resolution]
              properties:
                resolution: {type: string}
                lift_suspension: {type: boolean, default: false}
      responses:
        '200':
          description: Incident
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Incident'}
...
  /v1/doorstep/internal/admin/tickets/{id}/status:
    post:
      summary: Move a ticket
      operationId: adminSetTicketStatus
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:tickets.act
      x-lane: A6
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: {type: string, enum: [open, in_progress, resolved, closed]}
                note: {type: string}
      responses:
        '200':
          description: Ticket
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Ticket'}
...
  /v1/doorstep/internal/admin/ratings/{id}/hide:
    post:
      summary: Hide a rating (excluded from averages)
      operationId: adminHideRating
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:ratings.moderate
      x-lane: A6
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ReasonInput'}
      responses:
        '200':
          description: Rating
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Rating'}
...
components:
  securitySchemes:
    gatewayIdentity:
      type: apiKey
      in: header
      name: X-Verified-User-Id
      description: Stamped by the gateway from the verified JWT (with X-Internal-Service-Key). Never sent by clients directly.
    adminServiceToken:
      type: apiKey
      in: header
      name: X-Service-Authorization
      description: 'Bearer <ed25519 service token>: iss admin-service, aud doorstep, scope [one doorstep:* permission], act <admin user id>, exp <= 60 s.'
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: {type: string, format: uuid}
...
    CityQuery:
      name: city
      in: query
      required: true
      description: City code, e.g. HYD (case-insensitive)
      schema: {type: string, example: HYD}
...
  responses:
    Error:
      description: Error envelope (see x-doorstep-error-codes)
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
  schemas:
...
    Meta:
      type: object
      properties:
        request_id: {type: string}
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string, example: DOORSTEP_ADDON_INVALID}
            message: {type: string}
            details:
              type: object
              additionalProperties: true
        meta: {$ref: '#/components/schemas/Meta'}

    # ---------- enums ----------
...
    GenderRule:
      type: string
      enum: [any, female_pros_only, male_pros_only]
    ExtrasPolicy:
      type: string
      enum: [rate_card, catalogue_addons_only]
    BookingStatus:
      type: string
      enum: [pending_payment, confirmed, assigned, en_route, arrived, in_progress, awaiting_extras_payment, completed, cancelled, expired, customer_no_show, pro_no_show, pro_unavailable]
...
    CityRef:
      type: object
      required: [code, name]
      properties:
        code: {type: string, example: HYD}
        name: {type: string, example: Hyderabad}
    Catalogue:
      type: object
      required: [city, categories]
      properties:
        city: {$ref: '#/components/schemas/CityRef'}
        categories:
          type: array
          items: {$ref: '#/components/schemas/CategorySummary'}
    CategorySummary:
      type: object
      required: [id, slug, name, description, family, gender_rule, image_url, sort_order, service_count, starting_price_paise]
      properties:
        id: {type: string, format: uuid}
        slug: {type: string}
        name: {type: string}
        description: {type: string}
        family: {$ref: '#/components/schemas/Family'}
        gender_rule: {$ref: '#/components/schemas/GenderRule'}
        image_url: {type: string, nullable: true}
        sort_order: {type: integer}
        service_count: {type: integer, description: Active services with an active option}
        starting_price_paise: {type: integer, format: int64, nullable: true, description: "B1: the lowest bookable professional price of the category's options in the city now; null while no professional offers one"}
    CategoryPage:
      type: object
      required: [city, category, services]
      properties:
        city: {$ref: '#/components/schemas/CityRef'}
        category: {$ref: '#/components/schemas/CategorySummary'}
        services:
          type: array
          items: {$ref: '#/components/schemas/ServiceSummary'}
    ServiceSummary:
      type: object
      required: [id, category_id, slug, name, description, duration_minutes, image_url, starting_price_paise, suggested_price_paise]
      properties:
        id: {type: string, format: uuid}
        category_id: {type: string, format: uuid}
        slug: {type: string}
        name: {type: string}
        description: {type: string}
        duration_minutes: {type: integer}
        image_url: {type: string, nullable: true}
        starting_price_paise: {type: integer, format: int64, nullable: true, description: "B1: lowest bookable professional price of its options (null: nobody yet)"}
        suggested_price_paise: {type: integer, format: int64, nullable: true, description: "B1: lowest city suggested price of its options (informational, never charged); replaces starting_mrp_paise"}
    ServicePage:
      type: object
      required: [city, service]
      properties:
        city: {$ref: '#/components/schemas/CityRef'}
        service: {$ref: '#/components/schemas/ServiceDetail'}
    ServiceDetail:
      type: object
      required: [id, category, slug, name, description, duration_minutes, inclusions, exclusions, image_url, crew_size, rework_days, min_before_photos, min_after_photos, options, addon_groups]
      properties:
        id: {type: string, format: uuid}
        category:
          type: object
          required: [id, slug, name, family, gender_rule, extras_policy]
          properties:
            id: {type: string, format: uuid}
            slug: {type: string}
            name: {type: string}
            family: {$ref: '#/components/schemas/Family'}
            gender_rule: {$ref: '#/components/schemas/GenderRule'}
            extras_policy: {$ref: '#/components/schemas/ExtrasPolicy'}
        slug: {type: string}
        name: {type: string}
        description: {type: string}
        duration_minutes: {type: integer}
        inclusions: {type: array, items: {type: string}}
        exclusions: {type: array, items: {type: string}}
        image_url: {type: string, nullable: true}
        crew_size: {type: integer, description: Always 1 at launch (crew bookings switched off)}
        rework_days: {type: integer}
        min_before_photos: {type: integer}
        min_after_photos: {type: integer}
        options:
          type: array
          items: {$ref: '#/components/schemas/ServiceOption'}
        addon_groups:
          type: array
          items: {$ref: '#/components/schemas/AddonGroup'}
...
    AddonGroup:
      type: object
      required: [id, name, min_select, max_select, is_required, addons]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        min_select: {type: integer}
        max_select: {type: integer}
        is_required: {type: boolean}
        addons:
          type: array
          items: {$ref: '#/components/schemas/Addon'}
...
    ServiceabilityRequest:
      type: object
      required: [lat, lng]
      properties:
        lat: {type: number, format: double, minimum: -90, maximum: 90}
        lng: {type: number, format: double, minimum: -180, maximum: 180}
    Serviceability:
      type: object
      required: [serviceable, city, zone, reason]
      properties:
        serviceable: {type: boolean}
        city:
          allOf: [{$ref: '#/components/schemas/CityRef'}]
          nullable: true
        zone:
          type: object
          nullable: true
          required: [id, name]
          properties:
            id: {type: string, format: uuid}
            name: {type: string}
        reason:
          type: string
          nullable: true
          enum: [OUTSIDE_SERVICE_AREA, null]
    QuoteRequest:
      type: object
      required: [service_id, pro_id, option_id, lat, lng]
      properties:
        service_id: {type: string, format: uuid}
        pro_id: {type: string, format: uuid, description: "B1: the professional the customer picked (GET /services/{id}/professionals)"}
        option_id: {type: string, format: uuid}
        quantity: {type: integer, minimum: 1, default: 1}
        addons:
          type: array
          items:
            type: object
            required: [addon_id]
            properties:
              addon_id: {type: string, format: uuid}
        lat: {type: number, format: double}
        lng: {type: number, format: double}
    Quote:
      type: object
      required: [id, status, service_id, option_id, quantity, pro_id, city_code, zone_id, lines, total_paise, taxable_paise, tax_paise, prices_include_tax, tax_provisional, tax_note, duration_minutes, expires_at, created_at]
      properties:
        id: {type: string, format: uuid}
        status: {type: string, enum: [open, expired, consumed]}
        service_id: {type: string, format: uuid}
        option_id: {type: string, format: uuid}
        quantity: {type: integer}
        pro_id: {type: string, format: uuid, description: "B1: the picked professional; the quote is priced with their approved prices"}
        city_code: {type: string}
        zone_id: {type: string, format: uuid}
        lines:
          type: array
          items: {$ref: '#/components/schemas/QuoteLine'}
        total_paise: {type: integer, format: int64, description: What the customer pays (GST-inclusive)}
        taxable_paise: {type: integer, format: int64}
        tax_paise: {type: integer, format: int64}
        prices_include_tax: {type: boolean}
        tax_provisional:
          type: boolean
          description: true while the stub TaxComputer is in use (shared/gst Doorstep categories pending adviser sign-off)
        tax_note: {type: string}
        duration_minutes: {type: integer}
        expires_at: {type: string, format: date-time}
        created_at: {type: string, format: date-time}
    QuoteLine:
      type: object
      required: [kind, ref_id, price_id, name, unit, quantity, unit_price_paise, line_total_paise, taxable_paise, tax_paise, tax_rate_bps, gst_category, sac]
      properties:
        kind: {type: string, enum: [option, addon]}
        ref_id: {type: string, format: uuid}
        price_id: {type: string, format: uuid, description: "B1: the professional's doorstep.pro_service_prices row used (always an approved one)"}
        name: {type: string}
        unit: {type: string, enum: [per_job, per_hour, per_month]}
        quantity: {type: integer, description: Units (hours, months) for per_hour and per_month}
        unit_price_paise: {type: integer, format: int64}
        line_total_paise: {type: integer, format: int64}
        taxable_paise: {type: integer, format: int64}
        tax_paise: {type: integer, format: int64}
        tax_rate_bps: {type: integer}
        gst_category: {type: string, example: HOME_CLEANING_VIA_ECO}
        sac: {type: string}

    # ---------- customer (A3/A5) ----------
...
    Rating:
      type: object
      required: [id, booking_id, rater_kind, stars, tags, comment, hidden, created_at]
      properties:
        id: {type: string, format: uuid}
        booking_id: {type: string, format: uuid}
        rater_kind: {type: string, enum: [customer, pro]}
        stars: {type: integer}
        tags: {type: array, items: {type: string}}
        comment: {type: string, nullable: true}
        hidden: {type: boolean}
        created_at: {type: string, format: date-time}
...
    Incident:
      type: object
      required: [id, booking_id, raised_by_kind, kind, severity, status, description, pro_auto_suspended, created_at]
      properties:
        id: {type: string, format: uuid}
        booking_id: {type: string, format: uuid, nullable: true}
        raised_by_kind: {type: string, enum: [customer, pro, system, admin]}
        kind: {type: string, enum: [sos, safety, damage, harassment, theft, unsafe_exit, other]}
        severity: {type: string, enum: [low, medium, high, critical]}
        status: {type: string, enum: [open, acknowledged, resolved]}
        description: {type: string, nullable: true}
        pro_auto_suspended: {type: boolean}
        created_at: {type: string, format: date-time}
...
    SharedBookingView:
      type: object
      required: [status, locality, professional_first_name, slot_start, eta_minutes]
      properties:
        status: {$ref: '#/components/schemas/BookingStatus'}
        locality: {type: string}
        professional_first_name: {type: string, nullable: true}
        slot_start: {type: string, format: date-time}
        eta_minutes: {type: integer, nullable: true}
...
    Ticket:
      type: object
      required: [id, booking_id, category, subject, body, status, created_at, updated_at]
      properties:
        id: {type: string, format: uuid}
        booking_id: {type: string, format: uuid, nullable: true}
        category: {type: string}
        subject: {type: string}
        body: {type: string}
        status: {type: string, enum: [open, in_progress, resolved, closed]}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    TaxRegistration:
      type: object
      required: [pro_id, gstin, updated_at]
      properties:
        pro_id: {type: string, format: uuid}
        gstin: {type: string, nullable: true, description: 'Admin only. Never included in public professional cards or audit metadata.'}
        updated_at: {type: string, format: date-time}
...
    ReasonInput:
      type: object
      properties:
        reason: {type: string, maxLength: 1000}
...